// src/types/PreviewMedia.ts
export type PreviewMedia = {
    game_title?: string | null;

    movie_id?: number | null;
    youtube?: string | null;

    screen_shot_id_1?: number | null;
    path_full_1?: string | null;

    screen_shot_id_2?: number | null;
    path_full_2?: string | null;

    screen_shot_id_3?: number | null;
    path_full_3?: string | null;

    screen_shot_id_4?: number | null;
    path_full_4?: string | null;
};